:sparkles: feat: Store 예외처리
